2025-04-14

+++ umh-core/pkg/config/redpandaserviceconfig/normalizer_test.go
@@ -26,8 +26,8 @@ var _ = Describe("Redpanda YAML Normalizer", func() {
 			normalizer := NewNormalizer()
 
 			normalizedConfig := normalizer.NormalizeConfig(config)
-			Expect(normalizedConfig.Topic.DefaultTopicRetentionMs).To(Equal(0))
-			Expect(normalizedConfig.Topic.DefaultTopicRetentionBytes).To(Equal(0))
+			Expect(normalizedConfig.Topic.DefaultTopicRetentionMs).To(Equal(int64(0)))
+			Expect(normalizedConfig.Topic.DefaultTopicRetentionBytes).To(Equal(int64(0)))
 		})
 	})
 })
